+++ lib/statesman/machine.rb
@@ -2,7 +2,6 @@ require_relative "version"
 require_relative "exceptions"
 require_relative "guard"
 require_relative "callback"
-require_relative "event_transitions"
 require_relative "adapters/memory_transition"
 
 module Statesman
@@ -32,10 +31,6 @@ module Statesman
         @states ||= []
       end
 
-      def events
-        @events ||= {}
-      end
-
       def state(name, options = { initial: false })
         name = name.to_s
         if options[:initial]
@@ -43,10 +38,6 @@ module Statesman
           @initial_state = name
         end
         states << name
-      end
-
-      def event(name, &block)
-        EventTransitions.new(self, name, &block)
       end
 
       def successors
@@ -62,9 +53,9 @@ module Statesman
         }
       end
 
-      def transition(options = { from: nil, to: nil }, event = nil)
-        from = to_s_or_nil(options[:from])
-        to = array_to_s_or_nil(options[:to])
+      def transition(from: nil, to: nil)
+        from = to_s_or_nil(from)
+        to = array_to_s_or_nil(to)
 
         raise InvalidStateError, "No to states provided." if to.empty?
 
@@ -73,12 +64,6 @@ module Statesman
         ([from] + to).each { |state| validate_state(state) }
 
         successors[from] += to
-
-        if event
-          events[event] ||= {}
-          events[event][from] ||= []
-          events[event][from] += to
-        end
       end
 
       def before_transition(options = { from: nil, to: nil }, &block)
@@ -232,21 +217,6 @@ module Statesman
       true
     end
 
-    def trigger!(event_name, metadata = {})
-      transitions = self.class.events.fetch(event_name) do
-        raise Statesman::TransitionFailedError,
-              "Event #{event_name} not found"
-      end
-
-      new_state = transitions.fetch(current_state) do
-        raise Statesman::TransitionFailedError,
-              "State #{current_state} not found for Event #{event_name}"
-      end
-
-      transition_to!(new_state.first, metadata)
-      true
-    end
-
     def execute(phase, initial_state, new_state, transition)
       callbacks = callbacks_for(phase, from: initial_state, to: new_state)
       callbacks.each { |cb| cb.call(@object, transition) }
@@ -256,19 +226,6 @@ module Statesman
       self.transition_to!(new_state, metadata)
     rescue TransitionFailedError, GuardFailedError
       false
-    end
-
-    def trigger(event_name, metadata = {})
-      self.trigger!(event_name, metadata)
-    rescue TransitionFailedError, GuardFailedError
-      false
-    end
-
-    def available_events
-      state = current_state
-      self.class.events.select do |_, transitions|
-        transitions.key?(state)
-      end.map(&:first)
     end
 
     private